refactor(rocprofiler-compute): fold the pipeline lookup into code objects
The lookup lived in its own module under pc_sampling, though it has nothing to
do with sampling: it applies to every disassembled line. Move it next to the
other code-object disassembly helpers, its only consumer's existing import.

- Name it instruction_pipeline, which says what it returns and drops the alias
  the call site needed.
- Cover the operand form, s_waitcnt lgkmcnt(0), while moving the tests.
