+++ projects/rocprofiler-compute/src/pc_sampling/code_object_analysis.py
@@ -6,15 +6,26 @@
 Parse the per-process code-object info artifact emitted by the native PC
 sampling collector into a per-code-object instruction tree. The artifact holds
 the full disassembly of every loaded code object, including un-sampled ones.
+
+Also maps a disassembled instruction to the execution pipeline that runs it.
 """
 
+import functools
 import json
 from pathlib import Path
 from typing import Any, NamedTuple, Optional
 
+import yaml
+
+import config
 from utils.logger import console_warning
 
 CODE_OBJ_INFO_GLOB = "**/*_code_obj_info.json"
+
+INSTRUCTION_PIPELINES_FILE = "instruction_pipelines.yaml"
+# The C loader is an order of magnitude faster on a file this size, and ships
+# with PyYAML wherever libyaml is available.
+YAML_LOADER = getattr(yaml, "CSafeLoader", yaml.SafeLoader)
 
 
 class CodeObjectInstruction(NamedTuple):
@@ -77,6 +88,19 @@ def load_code_object_disassemblies(
     return disassemblies_per_pid
 
 
+def instruction_pipeline(instruction: Optional[str]) -> Optional[str]:
+    """Return the execution pipeline that runs a disassembled instruction.
+
+    The pipeline (VALU, MATRIX, SCALAR, and so on) is a property of the
+    instruction itself, so it applies to every disassembled line, not only the
+    offsets PC sampling landed on. Returns None for an instruction the table
+    does not hold, leaving the type unset rather than guessing.
+    """
+    if not instruction:
+        return None
+    return _instruction_pipelines().get(instruction.split(maxsplit=1)[0])
+
+
 def _parse_pid(json_path: Path) -> Optional[int]:
     """Extract the leading integer pid from a code-object info filename."""
     pid_text = json_path.name.split("_", 1)[0]
@@ -105,3 +129,31 @@ def _to_instruction(instruction: dict[str, Any]) -> CodeObjectInstruction:
         instruction=instruction.get("name"),
         source=instruction.get("comment"),
     )
+
+
+@functools.lru_cache(maxsize=1)
+def _instruction_pipelines() -> dict[str, str]:
+    """Load the generated table as a mnemonic-to-pipeline map, once per process.
+
+    The file groups mnemonics under their pipeline, which reads better than
+    17,000 unordered entries; the lookup wants the inverse. It is generated by
+    ``tools/instruction_pipeline_generator.py``.
+    """
+    path = (
+        config.rocprof_compute_home
+        / "rocprof_compute_soc"
+        / "analysis_configs"
+        / INSTRUCTION_PIPELINES_FILE
+    )
+    if not path.is_file():
+        console_warning(
+            f"Instruction pipeline table missing ({path}); "
+            "instruction types will be empty."
+        )
+        return {}
+    document = yaml.load(path.read_text(encoding="utf-8"), Loader=YAML_LOADER) or {}
+    return {
+        mnemonic: pipeline
+        for pipeline, mnemonics in (document.get("pipelines") or {}).items()
+        for mnemonic in mnemonics
+    }
