--- projects/rocprofiler-compute/src/pc_sampling/instruction_type.py
@@ -1,64 +1,0 @@
-# Copyright (c) Advanced Micro Devices, Inc.
-# SPDX-License-Identifier:  MIT
-
-"""Static instruction execution pipeline lookup.
-
-Map a disassembled instruction to the pipeline that runs it (VALU, MATRIX,
-SCALAR, and so on) using the generated mnemonic table. This is a property of
-the instruction itself, so it applies to every disassembled line, not only the
-offsets PC sampling landed on.
-
-The table is generated by ``tools/instruction_pipeline_generator.py``.
-"""
-
-import functools
-from typing import Optional
-
-import yaml
-
-import config
-from utils.logger import console_warning
-
-INSTRUCTION_PIPELINES_FILE = "instruction_pipelines.yaml"
-# The C loader is an order of magnitude faster on a file this size, and ships
-# with PyYAML wherever libyaml is available.
-YAML_LOADER = getattr(yaml, "CSafeLoader", yaml.SafeLoader)
-
-
-def classify(instruction: Optional[str]) -> Optional[str]:
-    """Return the execution pipeline of a disassembly line.
-
-    Returns None for a line whose mnemonic the table does not hold, which
-    leaves the instruction type unset rather than guessing a pipeline.
-    """
-    if not instruction:
-        return None
-    mnemonic = instruction.split(maxsplit=1)[0]
-    return _load_instruction_pipelines().get(mnemonic)
-
-
-@functools.lru_cache(maxsize=1)
-def _load_instruction_pipelines() -> dict[str, str]:
-    """Load the generated table as a mnemonic-to-pipeline map, once per process.
-
-    The file groups mnemonics under their pipeline, which reads better than
-    17,000 unordered entries; the lookup wants the inverse.
-    """
-    path = (
-        config.rocprof_compute_home
-        / "rocprof_compute_soc"
-        / "analysis_configs"
-        / INSTRUCTION_PIPELINES_FILE
-    )
-    if not path.is_file():
-        console_warning(
-            f"Instruction pipeline table missing ({path}); "
-            "instruction types will be empty."
-        )
-        return {}
-    document = yaml.load(path.read_text(encoding="utf-8"), Loader=YAML_LOADER) or {}
-    return {
-        mnemonic: pipeline
-        for pipeline, mnemonics in (document.get("pipelines") or {}).items()
-        for mnemonic in mnemonics
-    }
