+++ projects/rocprofiler-compute/tests/unit/pc_sampling/test_code_object_analysis.py
@@ -3,13 +3,18 @@
 
 import json
 from pathlib import Path
+from unittest.mock import patch
 
 import common
+import pytest
+import yaml
 
+from pc_sampling import code_object_analysis
 from pc_sampling.code_object_analysis import (
     CodeObjectDisassembly,
     CodeObjectInstruction,
     CodeObjectSymbol,
+    instruction_pipeline,
     load_code_object_disassemblies,
     parse_code_object_info,
 )
@@ -171,3 +176,64 @@ def test_load_skips_malformed_json():
         assert load_code_object_disassemblies(str(workload_dir)) == {}
     finally:
         common.clean_output_dir(True, str(workload_dir))
+
+
+# =============================================================================
+# Execution pipeline lookup
+# =============================================================================
+
+
+@pytest.fixture(autouse=True)
+def clear_table_cache():
+    """Drop the per-process table cache so each test loads it again."""
+    code_object_analysis._instruction_pipelines.cache_clear()
+    yield
+    code_object_analysis._instruction_pipelines.cache_clear()
+
+
+@pytest.fixture
+def pipeline_table(tmp_path):
+    """Point the loader at a small generated table."""
+    analysis_configs = tmp_path / "rocprof_compute_soc" / "analysis_configs"
+    analysis_configs.mkdir(parents=True)
+    (analysis_configs / code_object_analysis.INSTRUCTION_PIPELINES_FILE).write_text(
+        yaml.safe_dump({
+            "commit": "0" * 40,
+            "pipelines": {
+                "VALU": ["v_mov_b32_e32"],
+                "INTERNAL": ["s_waitcnt"],
+                "MATRIX": ["v_mfma_f32_16x16x16f16"],
+            },
+        }),
+        encoding="utf-8",
+    )
+    with patch.object(code_object_analysis.config, "rocprof_compute_home", tmp_path):
+        yield
+
+
+@pytest.mark.parametrize(
+    "instruction, expected",
+    [
+        ("v_mov_b32_e32 v1, 0", "VALU"),
+        ("s_waitcnt", "INTERNAL"),
+        ("s_waitcnt lgkmcnt(0)", "INTERNAL"),
+        ("v_mfma_f32_16x16x16f16 a[0:3], v0, v1, a[0:3]", "MATRIX"),
+        ("v_not_a_real_instruction v0", None),
+        (None, None),
+        ("", None),
+    ],
+)
+def test_instruction_pipeline_reads_the_leading_mnemonic(
+    pipeline_table, instruction, expected
+):
+    assert instruction_pipeline(instruction) == expected
+
+
+def test_instruction_pipeline_without_a_table_warns_once(tmp_path):
+    """A missing table leaves every type unset instead of failing analyze."""
+    with patch.object(code_object_analysis.config, "rocprof_compute_home", tmp_path):
+        with patch.object(code_object_analysis, "console_warning") as console_warning:
+            assert instruction_pipeline("v_mov_b32_e32 v1, 0") is None
+            assert instruction_pipeline("s_waitcnt") is None
+
+    console_warning.assert_called_once()
